+++ app/PersonioIntegration/Positions.php
@@ -195,8 +195,8 @@ class Positions {
 		 *
 		 * @since 3.0.0 Available since 3.0.0.
 		 *
-		 * @param array $query The configured query.
-		 * @param array $parameter_to_add The parameter to filter for.
+		 * @param array<string,mixed> $query The configured query.
+		 * @param array<string,mixed> $parameter_to_add The parameter to filter for.
 		 */
 		$query = apply_filters( 'personio_integration_positions_query', $query, $parameter_to_add );